# docs/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GabFlow API
  version: 0.4.0
  description: API principal para gestão de gabinetes municipais.
servers:
  - url: https://api.gabflow.example/api/v1
security:
  - cookieAuth: []
paths:
...
  /solicitacoes/{id}/encaminhamentos:
    post:
      summary: Encaminha uma solicitação para órgão externo
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201": { description: Encaminhamento criado }
  /encaminhamentos/{id}:
    patch:
      summary: Registra protocolo ou resposta do órgão externo
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Encaminhamento atualizado }
  /solicitacoes/{id}/reabrir:
    post:
      summary: Reabre uma solicitação encerrada mediante motivo
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Solicitação reaberta }
  /solicitacoes/{id}/chave-publica:
    post:
      summary: Gera e rotaciona a chave de acompanhamento público
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Chave gerada }
  /publico/solicitacoes/{protocolo}:
    get:
      summary: Consulta pública mínima da solicitação
      security: []
      parameters:
        - in: path
          name: protocolo
          required: true
          schema: { type: string }
        - in: query
          name: chave
          required: true
          schema: { type: string }
      responses:
        "200": { description: Estado público da solicitação }
        "403": { description: Chave inválida }
...
  /solicitacoes/{id}/tentativas-contato:
    post:
      summary: Registra tentativa de contato com o cidadão
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resultado]
              properties:
                canal:
                  type: string
                  enum: [WHATSAPP, TELEFONE, EMAIL, PRESENCIAL]
                destino: { type: string, maxLength: 254 }
                resultado:
                  type: string
                  enum: [REALIZADO, SEM_RESPOSTA, FALHOU, AGENDADO]
                observacoes: { type: string }
                justificativaCanal: { type: string }
                tentadaEm: { type: string, format: date-time }
                proximaTentativaEm: { type: string, format: date-time }
      responses:
        "201": { description: Tentativa registrada }
        "422": { description: Canal, destino ou resultado inválido }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...